Use isolates to fix multi threading issue
As I could not find a way to perform multi threading in Dart the UI was blocked at first. Then I found a dirty solution: waiting for some seconds hoping the rendering process would be done in the meantime. I was not satisfied by this solution but as this project is just meant to do some performance testing I could live with it at first. Then I found isolates which allowed me to do work without blocking the UI in Flutter.